popovers: Remove `hide_user_profile()` from `hide_all_except_sidebars()`.
`user-profile-modal` is shown using `overlays.open_overlay` which
disables mouse pointer events. The user can't click anywhere while a
modal is present, except to close it.

We use `hide_all_except_sidebars` and `hide_all` to hide popovers
while the user is interacting with other parts of the web app.
But since the user can't interact while a modal is present,
closing it manually is redundant.