delete_topic: Remove unnecessary `close_modal` call.
91b83cf09fd6485567de01e7dee1ad1c7c062447 migrated delete_topic modal
to confirm_dialog. confirm_dialog closes the modal after clicking on
the Confirm button. So manually closing the modal is redundant.

This went unnoticed since non-settings Bootstrap modals are hidden
instantly. By the time the request is successful, the modal is hidden,
and hiding it again doesn't raise an error.